ost-extractor/requirements.txt
UnityPy
ffmpeg-python
aioshutil
mutagen
pathvalidate
